Initial Commit
This is my first time working with/learning command line arguments but after toying around a bit it all makes sense now, now I just need to actually learn how to retrieve the DLL exports.

This is going to be a learning process but due to my extreme and random motivation to write this I am going to be dedicating my time to starting and finishing this.

I am sick and tired of manually writing my proxies, now this can generate them and work alongside my proxy template 'Proxy8'